requirements.txt
APScheduler==3.6.3
certifi==2026.1.4
cffi==2.0.0
charset-normalizer==3.4.4
cryptography==46.0.5
decorator==5.2.1
environs==14.5.0
google-api-core==2.30.0
google-api-python-client==2.151.0
google-auth==2.48.0
google-auth-httplib2==0.3.0
google-auth-oauthlib==1.2.4
googleapis-common-protos==1.72.0
gspread==6.2.1
httplib2==0.31.2
idna==3.11
marshmallow==4.2.2
oauthlib==3.3.1
proto-plus==1.27.1
protobuf==6.33.5
pyasn1==0.6.2
pyasn1_modules==0.4.2
pycparser==3.0
pyparsing==3.3.2
python-dotenv==1.0.1
python-telegram-bot==13.0
pytz==2025.2
requests==2.32.4
requests-oauthlib==2.0.0
rsa==4.9.1
six==1.17.0
tornado==6.5.4
tzlocal==5.3.1
uritemplate==4.2.0
urllib3==2.6.3